fix incorrect Hasher output
for sizes larger than a block+stripe (so 1088 bytes)
the Hasher would output the wrong value. this was due
to a bug caused by having reversed arguments to a copy
call. the test did not catch it because it too had a
bug where it was generating very aligned data bytes.

the reason why it does this copy is because the final
accumulation step always reads a stripe backwards from
the end, so if we just accumulated a block and then
need to finalize, it reads in the last stripe of bytes
again.

fixes #14