stop escaping input argument
the trick to dispatch to one of two functions by picking it
and putting it into a variable to allow the body to be inlined
has a problem: the compiler, seeing the virtual dispatch, is
not smart enough to notice that every option does not escape
the argument, and so it escapes the argument.

remove that and refactor some to gain some lost performance.
some smaller sizes are worse by <1ns, but some larger sizes
are better by up to 15%, and nothing escapes anymore.

fixes #19